Recognition on context
Element and Value classes added Basically. A BarChart has a list of
elements and an element has two values domain and range. Recognition in
each one of the areas is given by context, which means that in the
domain area system just recognize letters and numbers and in the range
just numbers. What is left is to map the domain, bar and range in the
elements list. Bars are also recognized.